pyproject.toml
[project]
name = "langchain-tool-and-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "groq>=0.37.1",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-core>=1.2.14",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchainhub>=0.1.21",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
